Console.CommandLine

Description

This file implements command line tools for startup of grill. In fact it currently is a general command line option parser.

| Author: Thorsten Rangwich. See file ../LICENSE for details.

Accessing command line options.

data Properties

Type storing command line properties. Properties are named by string and may occur more than once. The empty string marks command line arguments.

getFlag :: String -> Properties -> Bool

Retrieve a command line option without an argument.

getProp :: String -> Properties -> [String]

Get a property (list, if given more than once). Use getArguments to retrieve command line arguments instead of command line options.

getArguments :: Properties -> [String]

Retrieve the command line arguments from the properties.

Parsing command line

data PropertyType

Data constructor to specify type of command line option.

Constructors

Flag
Command line option is a flag without parameter.

OptOpt String String
Command line option is a flag with optional parameter description and default value.

ReqOpt String
Command line option is a flag with a required parameter description.

data Argument

Data constructor to specify one command line argument.

Constructors

Argument

Fields

shortNames :: String

String with all short option characters matching this option / flag.

longNames :: [String]

List of long option strings matching this option / flag.

description :: String

Description of this command line option.

propertyName :: String

Name of the property to identify this option within properties.

propertyType :: PropertyType

Type of this command line option.

parseOptions

Arguments

:: [Argument]
Options specification

-> [String]
Command line arguments

-> IO Properties
Parsed properties

Parse options and arguments into properties.

parseCmdLine

Arguments

:: [Argument]
Options speficication

-> IO Properties
Parsed object to be used for queries

Parse command line and return properties for further processing. further options for debugging are processed.