Data.Plain

Description

This file defines Plain, the base type used for the calculation trees.

Author: Thorsten Rangwich. See file ../LICENSE for details.

Base Data Types supported in Tree

data Plain

Plain data type containing some plain value or a composite result that should not be evaluated further.

Constructors

PlEmpty
Empty item

PlInt Int
Int value

PlFloat Float
Float value

PlString String
String value

PlError String
Error occured

Type Classes for Base Data Types

class Convert a where

Class to convert Plain into raw data type and vice versa.

Methods

get

Arguments

:: Plain
Input Plain

-> a
Raw data type for return

Convert Plain data type into its raw data type. Has to raise error in default implementation.

toPlain

Arguments

:: a
Input arbitrary data type.

-> Plain
Output - packaged in Plain.

Convert arbitrary data type into a packaged Plain. Default implementation returns PlError.

Tree Function Interface

type PlainFunction = [Plain] -> Plain

Function type for every tree function. Input takes a list of plains and returns one Plain value.

checkError

Arguments

:: Plain
Plain value to check for error

-> Bool
Return value - error contained in Plain?

Check Plain for error