Data.Sheet

Description

This file defines data structure in a workheet, build and access functions.

Author: Thorsten Rangwich. See file ../LICENSE for details.

Data types

data Sheet

(Hidden) type for raw sheet. May change in the future.

data Header

Type for sheet header. Just a properties map. May change in the future.

Sheet functions

createSheet

Arguments

:: String
Name of this sheet.

-> Sheet
Empty sheet created with defaults

Create a new sheet. Name is mandatory.

emptySheet

Arguments

:: Sheet
The default sheet returned.

Create new empty sheet.

buildSheet

Arguments

:: Header
Sheet header

-> [[FormulaTree]]
List of rows, rows itself lists of trees.

-> Sheet
Sheet structure.

Create a sheet from (usually parsed) contents

changeCell

Arguments

:: Address
Cell address

-> FormulaTree
Tree to add into cell

-> Sheet
Sheet to update

-> Sheet
Updated sheet FIXME: Set max updater for row and / or column to id, if nothing to do. Insert is quite more expensive then id and id will be sufficint in most cases.

Add / change a single cell to a raw sheet. May change and require certain conditions in the future.

getCell

Arguments

:: Sheet
The sheet.

-> Address
Cell address.

-> FormulaTree

Get content of single cell

numRows

Arguments

:: Sheet
The sheet

-> Coord
Number of rows

Number of rows in the sheet

numCols

Arguments

:: Coord
The row of interest.

-> Sheet
The sheet.

-> Coord
Number of columns in the row.

Number of columns of a particular row.

Header functions

createHeader

Arguments

:: String
Mandatory name of the sheet

-> Header
Default header returned

Create header with default settings. A name is mandatory.

requestHeader

Arguments

:: (SmallVersion, SmallVersion, SmallVersion)
Format of this sheet

-> (SmallVersion, SmallVersion, SmallVersion)
Level of calculation engine

-> String
Checksum -- FIXME: use a typedef for that

-> String
Sheet name

-> Either String Header
Error message or header

Request for a valid header for given versions.

addHeaderProperties

Arguments

:: [(String, Plain)]
Property name / value

-> Header
Old header

-> Header
updated header

Add properties to header.

Miscellaneous functions

defaultName :: String

Return the default name for a newly created sheet. This may change if there are more than one sheet bundled in a workbook as in that case there must be a counter to give any sheet a different name.