FormulaEngine.Evaluate

Description

This file implements evaluation of tree structure used for the formulas.

Author: Thorsten Rangwich. See file ../LICENSE for details.

Documentation

calcTree

Arguments

:: FormulaTree
Tree for evaluation.

-> Plain
Return value

Calculate a tree - without a sheet reference. Useful for tests only.

calcCell

Arguments

:: Sheet
The sheet to calculate a cell

-> Address
The cell address we are interested in

-> Plain
Plain value suitable for show

Calculate one cell in a sheet and return Plain value.

showTree

Arguments

:: FormulaTree
Tree to display

-> String

Show serialised format (edit format) of a cell (i.e. tree)