FormulaEngine.Parse

Description

This file implements parsing of both full sheets and single formula trees.

| Author: Thorsten Rangwich. See file ../LICENSE for details.

Documentation

compileTree

Arguments

:: String
Input string -- as given in a cell.

-> FormulaTree
Compiled formula tree.

Function to parse a string into a compiled FormulaTree.

compileSheet

Arguments

:: String
Input String. This may change to something else in the future.

-> Either String Sheet

Function to parse a whole sheet.

compileEditCell

Arguments

:: String
Input String.

-> Either String (Address, FormulaTree)

Function to compile an address and a new tree to put into address.