FormulaEngine.Registry

Description

This merges all formula functions into one great registry.

This is necessary to have everything available in the implementation of the tree evaluator.

Author: Thorsten Rangwich. See file ../LICENSE for details.

Documentation

resolve

Arguments

:: String
Name of the function

-> PlainFunction
The function found in the registry

Resolve a function name into a function operating in a tree