Tree.FormulaTree

Description

This file defines FormulaTree, the type used for modeling the calculation trees.

Author: Thorsten Rangwich. See file ../LICENSE for details.

Documentation

data FormulaTree

Compiled representation of a formula

Constructors

Raw Plain
Node is plain result value

Reference Address

Funcall NamedFunction [FormulaTree]
Node is a formula

TreeError TreeError
Node is evaluated to an error This makes sense for errors affecting the tree like circular references in contrast to plain value errors like div/0.

data TreeError

Type to mark error in tree

Constructors

NamedError String
Named error - this should be used

CircularReference
Circular reference in tree.

UnspecifiedError
Unspecified error - not useful, better add more specific errors

data NamedFunction

Wrapper to be used in the trees.

Constructors

NamedFunction

Fields

funcName :: String

funcCall :: PlainFunction