Version.Information

Description

This file provides and helps extract version information.

Author: Thorsten Rangwich. See file ../LICENSE for details.

Compile time and forever constants

Version information

versionString :: String

Version coded as string

formatString :: String

Maximum sheet format grill can handle

grillVersion :: (SmallVersion, SmallVersion, SmallVersion)

Grill's calculation version.

formatVersion :: (SmallVersion, SmallVersion, SmallVersion)

Sheet format version.

Validate version

checkGrill :: (SmallVersion, SmallVersion, SmallVersion) -> Maybe String

Check a given version if it is supported by the current calculation.

checkFormat :: (SmallVersion, SmallVersion, SmallVersion) -> Maybe String

Check a given version if it is supported by the current sheet formatter.

Parser to parse version from a string - only used by parsers

parseMagicBytes :: Parser ()

Parse magic bytes at the beginning of every grill sheet

parseFormat :: Parser (SmallVersion, SmallVersion, SmallVersion)

Parse sheet version in header.

parseVersion :: Parser (SmallVersion, SmallVersion, SmallVersion)

Parse calc engine version in header.

parseChecksum :: Parser String

Parse checksum in sheet header. Not yet implemented, always gives empty string currently.

parseEndOfHeader :: Parser ()

Parse header terminator.